+++ package.json
@@ -34,8 +34,8 @@
   ],
   "scripts": {
     "build": "yarn lint && tsc --project tsconfig.build.json",
-    "build:demo": "yarn lint && webpack --config webpack.build-demo.config.js",
-    "dev": "webpack-dev-server --hot --config webpack.dev.config.js",
+    "build:demo": "yarn lint && webpack --config webpack.build-demo.config.ts",
+    "dev": "webpack-dev-server --hot --config webpack.dev.config.ts",
     "lint": "eslint .",
     "test": "jest"
   },
@@ -50,6 +50,7 @@
     "@types/react-dom": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^7.3.1",
+    "@types/webpack": "^5.28.5",
     "eslint": "^8.56.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jest": "^26.1.5",
