+++ webpack.dev.config.ts
@@ -1,4 +1,5 @@
 import path from 'path';
+import webpack from 'webpack';
 import HtmlWebPackPlugin from 'html-webpack-plugin';
 import { fileURLToPath } from 'node:url'
 import ReactRefreshWebpackPlugin from '@pmmmwh/react-refresh-webpack-plugin';
@@ -7,7 +8,7 @@ import ESLintPlugin from 'eslint-webpack-plugin';
 const filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(filename)
 
-export default {
+const config: webpack.Configuration = {
   devtool: 'inline-source-map',
   mode: 'development',
   entry: path.resolve(dirname, 'demo/index.tsx'),
@@ -22,11 +23,11 @@ export default {
         loader: 'ts-loader',
         options: {
           transpileOnly: true,
+          compilerOptions: {
+            module: "ESNext",
+            moduleResolution: "node",
+          }
         },
-        compilerOptions: {
-          module: "ESNext",
-          moduleResolution: "node",
-        }
       }
     ],
   },
@@ -55,3 +56,5 @@ export default {
     host: '0.0.0.0',
   },
 };
+
+export default config;
